This is a fix for a problem Juerg Schmidli reported when trying to access a NetCDF variable with a large (> 30,000,000) dimension size from PyNIO. Code in the NetCDF module calculated the number of elements to be requested (considering the stride) by casting to float before dividing. This was losing enough precision that an incorrect request was made to the NetCDF library. I repaired this by casting to double instead. I simplified the expressions a bit as well, but I did not change the basic algorithm.
I made similar changes to other places I found in NCL that had similar code.
e


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9975 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/NclHDF.c
@@ -1,5 +1,5 @@
 /*
- *      $Id: NclHDF.c,v 1.25 2007-06-21 01:35:32 dbrown Exp $
+ *      $Id: NclHDF.c,v 1.26 2007-12-21 00:49:23 dbrown Exp $
  */
 /************************************************************************
 *									*
@@ -1253,7 +1253,7 @@ void* storage;
 	while(stepvl != NULL) {
 		if(stepvl->var_inq->name == thevar) {
 			for(i= 0; i< stepvl->var_inq->n_dims; i++) {
-				count[i] = (int)floor((float)(finish[i] - start[i])/(float)stride[i]) + 1;
+				count[i] = (int)floor((finish[i] - start[i])/(double)stride[i]) + 1;
 				n_elem *= count[i];
 				if(stride[i] != 1) {
 					no_stride = 0;
@@ -1476,7 +1476,7 @@ long *stride;
 		while(stepvl != NULL) {
 			if(stepvl->var_inq->name == thevar) {
 				for(i= 0; i< stepvl->var_inq->n_dims; i++) {
-					count[i] = (int)floor((float)(finish[i] - start[i])/(float)stride[i]) + 1;
+					count[i] = (int)floor((finish[i] - start[i])/(double)stride[i]) + 1;
 					n_elem *= count[i];
 					if(stride[i] != 1) {
 						no_stride = 0;
